refactor SQ/eval to live in core/eval and separate backend-specific and backend-neutral code
