+++ SeQuant/core/export/export.hpp
@@ -2,7 +2,7 @@
 #define SEQUANT_CORE_EXPORT_EXPORT_HPP
 
 #include <SeQuant/core/container.hpp>
-#include <SeQuant/core/eval_expr.hpp>
+#include <SeQuant/core/eval/eval_expr.hpp>
 #include <SeQuant/core/export/compute_selection.hpp>
 #include <SeQuant/core/export/context.hpp>
 #include <SeQuant/core/export/export_expr.hpp>
