+++ SeQuant/core/optimize/common_subexpression_elimination.hpp
@@ -3,8 +3,8 @@
 
 #include <SeQuant/core/binary_node.hpp>
 #include <SeQuant/core/container.hpp>
-#include <SeQuant/core/eval_expr.hpp>
-#include <SeQuant/core/eval_node.hpp>
+#include <SeQuant/core/eval/eval_expr.hpp>
+#include <SeQuant/core/eval/eval_node.hpp>
 #include <SeQuant/core/expr.hpp>
 #include <SeQuant/core/hash.hpp>
 #include <SeQuant/core/utility/macros.hpp>
