+++ cmake/sequant-config.cmake.in
@@ -28,6 +28,16 @@ if(SEQUANT_HAS_TILEDARRAY AND NOT TARGET tiledarray)
   find_dependency(TiledArray CONFIG QUIET REQUIRED COMPONENTS tiledarray PATHS ${TiledArray_DIR} NO_DEFAULT_PATH)
 endif()
 
+set(SEQUANT_HAS_BTAS @SEQUANT_HAS_BTAS@)
+if(SEQUANT_HAS_BTAS AND NOT TARGET BTAS::BTAS)
+  set(BTAS_CONFIG @BTAS_CONFIG@)
+  if (NOT BTAS_CONFIG OR NOT EXISTS ${BTAS_CONFIG})
+    message(FATAL_ERROR "Expected BTAS config file at ${BTAS_CONFIG}; directory moved since SeQuant configuration?")
+  endif()
+  get_filename_component(BTAS_DIR ${BTAS_CONFIG} DIRECTORY)
+  find_dependency(BTAS CONFIG QUIET REQUIRED PATHS ${BTAS_DIR} NO_DEFAULT_PATH)
+endif()
+
 set(SEQUANT_HAS_EIGEN @SEQUANT_HAS_EIGEN@)
 if (NOT TARGET Eigen3::Eigen AND SEQUANT_HAS_EIGEN)
   if (TARGET TiledArray_Eigen)
