+++ tests/integration/eval/CMakeLists.txt
@@ -1,6 +1,3 @@
-option(SEQUANT_EVAL_TESTS "Enable building of evaluation tests (if true, will look for and/or build TiledArray)" OFF)
-add_feature_info(EVAL_TESTS SEQUANT_EVAL_TESTS "Build evaluation tests (if true, will look for and/or build TiledArray)")
-
 # uccf12 example moved to MPQC
 
 add_library(eval_shared STATIC EXCLUDE_FROM_ALL
@@ -17,15 +14,13 @@ add_library(eval_shared STATIC EXCLUDE_FROM_ALL
 target_link_libraries(eval_shared PUBLIC SeQuant)
 target_include_directories(eval_shared PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}")
 
-if (TARGET tiledarray)
-    set(example5 eval_ta)
+if (SEQUANT_HAS_TILEDARRAY)
     add_executable(eval_ta ${BUILD_BY_DEFAULT}
             "ta/data_world_ta.hpp"
             "ta/scf_ta.hpp"
             "ta/main.cpp"
     )
     target_link_libraries(eval_ta PRIVATE eval_shared tiledarray)
-    target_compile_definitions(eval_ta PRIVATE SEQUANT_HAS_TILEDARRAY)
 
     set(test_name "sequant/integration/eval_ta")
     add_test(
@@ -35,14 +30,15 @@ if (TARGET tiledarray)
     )
 
     build_test_as_needed(eval_ta "${test_name}" test_name)
+endif (SEQUANT_HAS_TILEDARRAY)
 
+if (SEQUANT_HAS_BTAS)
     add_executable(eval_btas ${BUILD_BY_DEFAULT}
             "btas/data_world_btas.hpp"
             "btas/scf_btas.hpp"
             "btas/main.cpp"
     )
-    target_include_directories(eval_btas PRIVATE ${BTAS_SOURCE_DIR})
-    target_link_libraries(eval_btas PRIVATE eval_shared)
+    target_link_libraries(eval_btas PRIVATE eval_shared BTAS::BTAS)
 
     set(test_name "sequant/integration/eval_btas")
     add_test(
@@ -52,4 +48,4 @@ if (TARGET tiledarray)
     )
 
     build_test_as_needed(eval_btas "${test_name}" test_name)
-endif (TARGET tiledarray)
+endif (SEQUANT_HAS_BTAS)
