+++ tests/integration/eval/btas/data_world_btas.hpp
@@ -10,6 +10,8 @@
 
 #include <btas/btas.h>
 #include <SeQuant/core/container.hpp>
+#include <SeQuant/core/eval/backends/btas/eval_expr.hpp>
+#include <SeQuant/core/eval/backends/btas/result.hpp>
 #include <SeQuant/core/eval/eval.hpp>
 #include <SeQuant/core/expr.hpp>
 #include <SeQuant/core/utility/macros.hpp>
