+++ tests/integration/eval/btas/scf_btas.hpp
@@ -10,11 +10,12 @@
 #include <data_info.hpp>
 #include <scf.hpp>
 
+#include <SeQuant/core/eval/backends/btas/eval_expr.hpp>
 #include <SeQuant/core/eval/cache_manager.hpp>
 #include <SeQuant/core/eval/eval.hpp>
 
 #include <SeQuant/core/container.hpp>
-#include <SeQuant/core/eval_node.hpp>
+#include <SeQuant/core/eval/eval_node.hpp>
 #include <SeQuant/core/parse.hpp>
 #include <SeQuant/core/utility/macros.hpp>
