+++ tests/integration/eval/eval_utils.hpp
@@ -6,7 +6,7 @@
 #define SEQUANT_EVAL_EVAL_UTILS_HPP
 
 #include <SeQuant/core/container.hpp>
-#include <SeQuant/core/eval_node.hpp>
+#include <SeQuant/core/eval/eval_node.hpp>
 #include <SeQuant/core/utility/macros.hpp>
 #include <chrono>
 #include <fstream>
