+++ tests/integration/eval/ta/data_world_ta.hpp
@@ -9,7 +9,7 @@
 #include <eval_utils.hpp>
 
 #include <SeQuant/core/container.hpp>
-#include <SeQuant/core/eval/result.hpp>
+#include <SeQuant/core/eval/backends/tiledarray/result.hpp>
 #include <SeQuant/core/expr.hpp>
 #include <SeQuant/core/utility/macros.hpp>
