+++ tests/integration/eval/ta/scf_ta.hpp
@@ -6,6 +6,7 @@
 #define SEQUANT_EVAL_SCF_TA_HPP
 
 #include <SeQuant/core/container.hpp>
+#include <SeQuant/core/eval/backends/tiledarray/eval_expr.hpp>
 #include <SeQuant/core/eval/cache_manager.hpp>
 #include <SeQuant/core/eval/eval.hpp>
 #include <SeQuant/core/parse.hpp>
